add event store interfaces, refactor InMemoryEventStore
