--- src/Adapter/InMemoryAdapter.php
@@ -1,257 +1,0 @@
-<?php
-/**
- * This file is part of the prooph/event-store.
- * (c) 2014-2016 prooph software GmbH <contact@prooph.de>
- * (c) 2015-2016 Sascha-Oliver Prolic <saschaprolic@googlemail.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace Prooph\EventStore\Adapter;
-
-use AppendIterator;
-use ArrayIterator;
-use Iterator;
-use Prooph\EventStore\Exception\StreamNotFoundException;
-use Prooph\EventStore\Metadata\MetadataMatcher;
-use Prooph\EventStore\Metadata\Operator;
-use Prooph\EventStore\Stream\Stream;
-use Prooph\EventStore\Stream\StreamName;
-
-/**
- * Class InMemoryAdapter
- *
- * @package Prooph\EventStore\Adapter
- * @author Alexander Miertsch <contact@prooph.de>
- */
-class InMemoryAdapter implements Adapter
-{
-    /**
-     * @var array
-     */
-    protected $streams;
-
-    public function hasStream(StreamName $streamName): bool
-    {
-        return isset($this->streams[$streamName->toString()]);
-    }
-
-    public function fetchStreamMetadata(StreamName $streamName): ?array
-    {
-        if ($this->hasStream($streamName)) {
-            return $this->streams[$streamName->toString()]['metadata'];
-        }
-
-        return null;
-    }
-
-    public function create(Stream $stream): void
-    {
-        $streamEvents = $stream->streamEvents();
-        $streamEvents->rewind();
-
-        $this->streams[$stream->streamName()->toString()]['events'] = $streamEvents;
-        $this->streams[$stream->streamName()->toString()]['metadata'] = $stream->metadata();
-    }
-
-    /**
-     * @throws StreamNotFoundException
-     */
-    public function appendTo(StreamName $streamName, Iterator $domainEvents): void
-    {
-        if (! $this->hasStream($streamName)) {
-            throw StreamNotFoundException::with($streamName);
-        }
-
-        $it = new AppendIterator();
-        $it->append($this->streams[$streamName->toString()]['events']);
-        $it->append($domainEvents);
-
-        $this->streams[$streamName->toString()]['events'] = $it;
-    }
-
-    public function load(
-        StreamName $streamName,
-        int $fromNumber = 1,
-        int $count = null
-    ): ?Stream {
-        if (! $this->hasStream($streamName)) {
-            return null;
-        }
-
-        $filteredEvents = [];
-
-        foreach ($this->streams[$streamName->toString()]['events'] as $streamEvent) {
-            if ((null === $count
-                    && $streamEvent->version() >= $fromNumber
-                ) || (null !== $count
-                    && $streamEvent->version() >= $fromNumber
-                    && $streamEvent->version() <= ($fromNumber + $count - 1)
-                )
-            ) {
-                $filteredEvents[] = $streamEvent;
-            }
-        }
-
-        return new Stream(
-            $streamName,
-            new \ArrayIterator($filteredEvents),
-            $this->streams[$streamName->toString()]['metadata']
-        );
-    }
-
-    public function loadReverse(
-        StreamName $streamName,
-        int $fromNumber = PHP_INT_MAX,
-        int $count = null
-    ): ?Stream {
-        if (! $this->hasStream($streamName)) {
-            return null;
-        }
-
-        $streamEvents = $this->streams[$streamName->toString()]['events'];
-
-        $filteredEvents = [];
-
-        foreach ($streamEvents as $streamEvent) {
-            if ((null === $count
-                    && $streamEvent->version() <= $fromNumber
-                )
-                || (null !== $count
-                    && $streamEvent->version() <= $fromNumber
-                    && $streamEvent->version() >= ($fromNumber - $count + 1)
-                )
-            ) {
-                $filteredEvents[] = $streamEvent;
-            }
-        }
-
-        $filteredEvents = array_reverse($filteredEvents);
-
-        return new Stream(
-            $streamName,
-            new \ArrayIterator($filteredEvents),
-            $this->streams[$streamName->toString()]['metadata']
-        );
-    }
-
-    public function loadEvents(
-        StreamName $streamName,
-        int $fromNumber = 1,
-        int $count = null,
-        MetadataMatcher $metadataMatcher = null
-    ): Iterator {
-        if (! $this->hasStream($streamName)) {
-            return new ArrayIterator();
-        }
-
-        if (null === $metadataMatcher) {
-            $metadataMatcher = new MetadataMatcher();
-        }
-
-        $streamEvents = [];
-
-        foreach ($this->streams[$streamName->toString()]['events'] as $streamEvent) {
-            if ($this->matchesMetadata($metadataMatcher, $streamEvent->metadata())
-                && ((null === $count
-                        && $streamEvent->version() >= $fromNumber
-                    ) || (null !== $count
-                        && $streamEvent->version() < ($fromNumber + $count - 1)
-                    )
-                )
-            ) {
-                $streamEvents[] = $streamEvent;
-            }
-        }
-
-        return new ArrayIterator($streamEvents);
-    }
-
-    public function loadEventsReverse(
-        StreamName $streamName,
-        int $fromNumber = PHP_INT_MAX,
-        int $count = null,
-        MetadataMatcher $metadataMatcher = null
-    ): Iterator {
-        if (! $this->hasStream($streamName)) {
-            return new ArrayIterator();
-        }
-
-        if (null === $metadataMatcher) {
-            $metadataMatcher = new MetadataMatcher();
-        }
-
-        $streamEvents = [];
-
-        foreach ($this->streams[$streamName->toString()]['events'] as $index => $streamEvent) {
-            if ($this->matchesMetadata($metadataMatcher, $streamEvent->metadata())
-                && ((null === $count
-                        && $streamEvent->version() <= $fromNumber
-                    ) || (null !== $count
-                        && $streamEvent->version() <= ($fromNumber - $count + 1)
-                    )
-                )
-            ) {
-                $streamEvents[] = $streamEvent;
-            }
-        }
-
-        $streamEvents = array_reverse($streamEvents);
-
-        return new ArrayIterator($streamEvents);
-    }
-
-    private function matchesMetadata(MetadataMatcher $metadataMatcher, array $metadata): bool
-    {
-        foreach ($metadataMatcher->data() as $match) {
-            $key = $match['key'];
-
-            if (! isset($metadata[$key])) {
-                return false;
-            }
-
-            $operator = $match['operator'];
-            $expected = $match['value'];
-
-            switch ($operator) {
-                case Operator::EQUALS():
-                    if ($metadata[$key] !== $expected) {
-                        return false;
-                    }
-                    break;
-                case Operator::GREATER_THAN():
-                    if (! ($metadata[$key] > $expected)) {
-                        return false;
-                    }
-                    break;
-                case Operator::GREATER_THAN_EQUALS():
-                    if (! ($metadata[$key] >= $expected)) {
-                        return false;
-                    };
-                    break;
-                case Operator::LOWER_THAN():
-                    if (! ($metadata[$key] < $expected)) {
-                        return false;
-                    }
-                    break;
-                case Operator::LOWER_THAN_EQUALS():
-                    if (! ($metadata[$key] <= $expected)) {
-                        return false;
-                    }
-                    break;
-                case Operator::NOT_EQUALS():
-                    if ($metadata[$key] === $expected) {
-                        return false;
-                    }
-                    break;
-                default:
-                    throw new \UnexpectedValueException('Unknown operator found');
-            }
-        }
-
-        return true;
-    }
-}
