--- src/Adapter/PayloadSerializer.php
@@ -1,26 +1,0 @@
-<?php
-/**
- * This file is part of the prooph/event-store.
- * (c) 2014-2016 prooph software GmbH <contact@prooph.de>
- * (c) 2015-2016 Sascha-Oliver Prolic <saschaprolic@googlemail.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace Prooph\EventStore\Adapter;
-
-/**
- * Interface PayloadSerializer
- *
- * @package Prooph\EventStore\Adapter
- * @author Alexander Miertsch <contact@prooph.de>
- */
-interface PayloadSerializer
-{
-    public function serializePayload(array $payload): string;
-
-    public function unserializePayload(string $payloadStr): array;
-}
